repo cleanup, split code into 3 branches
